Add support for CSS Device Adaptation
Viewport sizing is currently declared in a meta tag which is not a CSS
standard. IE10 and the latest Opera browsers now support the CSS
standard declaration (http://dev.w3.org/csswg/css-device-adapt/) so this
is to add support for them.

See http://timkadlec.com/2012/10/ie10-snap-mode-and-responsive-design/
for details of the IE10 implementation

Support in other browsers is currently in progress. The following
feature requests for Webkit, Chromium & Firefox are listed as these
bugs:

https://bugs.webkit.org/show_bug.cgi?id=95959
https://bugs.webkit.org/show_bug.cgi?id=95959
https://bugzilla.mozilla.org/show_bug.cgi?id=747754