Remove rem value rules
We were never really benifiting from using rems as we never changed our
base font-size.

Using rems was causing issues with firefox due to the way it has
implemented minimum font size. In firefox the minimum font-size ensures
no element can have a font-size lower than the specified size. This was
causing the html element to have a font-size bigger than the 10px we
were expecting. This caused some fonts on GOV.UK to be huge.