Use variables for font-sizes rather than arguments
Calls to Sass mixins need to have positional arguments before named
ones. This was causing all the calls to the font-generator to fail.
Rather than removing the named arguments, because that would use magic
number arguments which are confusing, store the values to pass to the
generator in varaibles and only pass in positional arguments.