Make Tracker.js more liberal in what it accepts
Tracker.js currently expects config.universalId and config.classicId to
be set, meaning to use it we have to have both classic and universal
analytics.

This change follows on from the previous commit to normalise the
interfaces to google-analytics-universal-tracker and
google-analytics-classic-tracker meaning that this code can treat
both trackers the same without having to know what type of tracker
it is (classic or universal).

This means that we can start switching off classic analytics in our
apps by just removing one line of config.

This commit switches over the code in tracker.js to handle trackers
generically, and adds a test to ensure that the Tracker initialiser
doesn't blow up when only one tracking code is passed in.