Add support for sending page option when tracking events
This adds support for sending the `page` option to `ga()` when tracking
events[1].

We're trying to use events to track when people reach an outcome in Smart
Answers[2]. The current set-up is giving us inaccurate data because of the use
of JavaScript to partially replace pages as a user navigates the Smart Answer:
The majority of the events are being recorded against the start page of the
Smart Answer (e.g. /student-finance-forms/y) instead of the outcome page.

I've got an open pull request in Smart Answers that switches us to use `ga()`
directly so that I can test whether sending the `page` option actually gives us
the data we're interested in[3]. Assuming it does then I think the right thing
to do is add support for `page` in this project.

[1]: https://developers.google.com/analytics/devguides/collection/analyticsjs/events#implementation
[2]: https://github.com/alphagov/smart-answers/commit/5d09931ce02c65ce1ca255f74960e1ee4ec2681e
[3]: https://github.com/alphagov/smart-answers/pull/1787