Add external link tracking plugin
The simplest approach to external link tracking, using Google Analytics
`sendBeacon` support to guarantee the event is captured when supported.

When not supported, the event will fire and a proportion of those events
will make it to analytics. The dataset will be incomplete.

Add support for setting the transport option to Universal Tracker