Add download link tracker plugin
The simplest approach to download tracking, using Google Analytics
`sendBeacon` support to guarantee the pageview is captured when
supported.

When not supported, the pageview will fire and a proportion of those
events will make it to analytics. The dataset will be incomplete.

Includes adding support for setting the transport option when tracking
page views.