Specify an explicit ruby version
The lack of an explicit ruby version was causing builds to fail in
GOV.UK CI, as it fell back to a system ruby without `bundle`.

We were setting an explict version just for travis, which i've moved
to `.ruby-version`, as it's more widely recognised and still used by
travis.