Find and auto-start modules from markup
Find and auto-start modules specified using the `data-module=""`
pattern in markup.

Each module must register a constructor with GOVUK.Modules and
have a `start` method. The start method receives the element that
the data attribute is defined on.

This makes it easier to limit modules to their containers and to
start modules within certain elements We can also remove inline
script tags and stop worrying about instantiation.

Based on the govuk_admin_template:
https://github.com/alphagov/govuk_admin_template/pull/3