Check module HTML explicitly in tests
The object conversion when running tests on CI revealed a deficiency in
the existing module assertions that attempted to use `toMatch` on HTML
objects.

Instead assert that the HTML we expect is within the jQuery object.

This unveiled a failing test which has also been fixed. The `push`
method for when a container is also a module should actually be `add`.