Fix scss lint offences
Fixes are mainly trimming trailing whitespace, adding spaces and newlines.

Warnings which are disabled:

_css3.scss
// scss-lint:disable Indentation SpaceAfterPropertyColon
These are caused by the tabbed indentation, which is intentional.

_font-stack.scss
_typography.scss
// scss-lint:disable NameFormat
Ignore errors for font variable names

_media-player.scss
// scss-lint:disable ColorVariable
Disable for now, but check that the colour orange here is correct.
This is a focus style, so should use $focus-colour.