Set duration (in days) for AB Test cookie.
Here, I have added a new option (cookieDuration), this option help to set the cookie expiration duration in days.

When this option isn't supplied the the default cookie duration is set to expire in 30 days.

The motivation here is, by reducing the cookie expiration will allow returning users to be assigned
to different cohorts during A/B tests.

According to the performance analyst, A/B tests performance is measured using sessions and with a default 30 day expiration,
they will only see the one variant for that period.  Being able to reduce this increases the likelihood of users seeing different variants
across their sessions.