Run handleRadioContent only on controlling radios
This reduces the radios we are trying to hide the
content for to only be those which are controls.

This covers the following use cases:

1. If the control sent in is not a control

All content controlled by radios in this group
will be hidden.

2. If the control sent in is a control

All content controlled by radios in this group
will be hidden and the content for the control
will be shown.