Only bind events to controlling selection buttons
At the moment the ShowHideContent module binds
events to all radios and checkboxes in
block-labels. This mean a user clicking on any
block-label in the page will trigger code to
show/hide its related content, even if there is
none.

This commit limits the block-labels that trigger
the code to show/hide content related to it to
those that have that relationship.