Fix a weird stick at top when scrolling bug
For reasons unknown, getting the width of the stuck element by calling $el.width() _after_ adding the shim to the document causes Chrome 56 to scroll down by the height of the shim, effectively ‘jumping’ the user down the document.

This behaviour does not occur in other browsers, nor did it occur in Chrome 55.

Somehow, this fixes it.

¯\_(ツ)_/¯