Perform Travis deployment setup only when deploying
Steps in `before_install` are run for all builds, not just deploys.

For branches/PRs from forks of this repo the install step fails as the
user will not have access to the ENV vars used to decrypt files, like
`encrypted_909ac1036a94_key`. This was caught in #383, where a forked
PR build consistently fails on the `openssl` step.

Steps specified in `before_deploy` are only run if a deployment is
triggered [^before_deploy], which will only happen on `master`, which
can't be triggered by a forked user.

Also adds a couple of inline comments to help explain the steps, which
are a little obtuse at first glance.

[^before_deploy]: [Travis `before_deploy` documentation](https://docs.travis-ci.com/user/deployment/npm/#Running-commands-before-and-after-deploy)