Fix js error causing missing accordions in IE7
Accessing individual characters in strings as if they are arrays is an ECMAScript 5 feature which was only introduced in IE8. So IE 7 and below (and IE11 when in compatibility mode) are erroring because of this.

Notably this is leaving accordions in a state where the original page content has been removed (because they use the js-hidden class), but the accordion fails to initialise, leaving the user with no way to navigate further.

We can fix this by using `charAt`[1] as recommended by MDN [2]

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/charAt
[2]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String#Character_access