Add package.lock
This is an artefact of npm 5+, which is an equivalent of the
`Gemfile.lock` in Ruby-land, in that it captures the exact versions
(with a version number and a hash for integrity checking) of all
sub-dependencies.

It’s recommended that this is checked in to source control, so that you
can be sure that the exact same dependencies are used by all users
(e.g. developers, continuous integration servers, staging, production,
etc).

See http://blog.npmjs.org/post/161081169345/v500 for details.