Add custom dimension to be joined on by GA upload
This change is meant to duplicate the url information in a custom dimension. The url is usually sent in the event action. However, we need this info to also be present in a custom dimension.
The reason for this is that we are doing a join between the click hit data and the information we upload via a Google Analytics Custom Upload. The custom upload is called "External Link Status" and needs to join existing data strictly on a custom dimension.