Treat `Arguments` like `Array` instances when stripping PII
Differences in JS runtimes means that we can't always iterate over an
`Arguments` instance's properties with a `for (var prop in object) {
... }` loop like we can for other object types and we have to iterate
over the contents like we do with an array with a `for (var i = 0, l =
array.length; i < l; i++) { ... }` loop.  If we don't iterate this way
then for those runtimes that don't allow it we don't run the `stripPII`
function on the contents of the `Arguments` instance and all is for
naught!

It seems that most current browsers do let you iterate over an `Argument`
object using properties, but the phantomjs runtime that GOV.UK uses in
the static application to run tests doesn't let you do this.  It's likely
this is an old implementation, but it's probably based on some browser
behaviour, so it means that some older browsers may also not allow this.

Iterating by numerical index works in all browsers and runtimes so it's
safer to do this.