Remove deprecation message and replace with notice
On GOV.UK sites we see this console alert a heck of a lot particularly
in test environments (see example below from [publishing-e2e-tests][1])
- I dread to imagine how many bytes we're using to store it in various
CI logs.

Since this deprecation depends on a third party library that GOV.UK is
not fully utilising it doesn't seem to be particularly helpful as the
only way to get rid of it is to switch to using GOV.UK elements.
Therefore I think it's more helpful if this is switched from being a
deprecation to be a notice in the source code.

A nicer alternative would be if the condition that checked for console
also had means to check if GOV.UK elements was being used.

End to end tests example:

```
17:45:03 Upload attachments on Travel Advice Publisher
17:45:04 Deprecation warning: Custom radio buttons and checkboxes (released in GOV.UK Elements 3.0.0) no longer require this JavaScript.
17:45:04 Deprecation warning: Custom radio buttons and checkboxes (released in GOV.UK Elements 3.0.0) no longer require this JavaScript.
17:45:04 Deprecation warning: Custom radio buttons and checkboxes (released in GOV.UK Elements 3.0.0) no longer require this JavaScript.
17:45:05 Deprecation warning: Custom radio buttons and checkboxes (released in GOV.UK Elements 3.0.0) no longer require this JavaScript.
17:45:05 Deprecation warning: Custom radio buttons and checkboxes (released in GOV.UK Elements 3.0.0) no longer require this JavaScript.
17:45:05 Deprecation warning: Custom radio buttons and checkboxes (released in GOV.UK Elements 3.0.0) no longer require this JavaScript.
17:45:05 Deprecation warning: Custom radio buttons and checkboxes (released in GOV.UK Elements 3.0.0) no longer require this JavaScript.
17:45:05 Deprecation warning: Custom radio buttons and checkboxes (released in GOV.UK Elements 3.0.0) no longer require this JavaScript.
17:45:05 Deprecation warning: Custom radio buttons and checkboxes (released in GOV.UK Elements 3.0.0) no longer require this JavaScript.
17:45:05   Publishing a manual
```

[1]: https://github.com/alphagov/publishing-e2e-tests