Support cross domain event tracking
Adds the ability to specify a trackerName option when calling
trackEvent, this can be used to refer to a linked tracker (ie. a
tracker used for cross domain analytics) which allows events to
be sent to other departmental analytics accounts.

Adds the ability to override the default behaviour of sending a
pageview when calling addLinkedTrackerDomain. This is useful when
only events are required for cross domain tracking.