Bump to version 9.0.1
I'm considering the removal of ruby-sass to be a patch release as I
don't believe it actually provided consumers with any help and thus is
essentially a bug.

The other option I considered was a major release as the lack of
dependency may have indicated a breaking change, however I can't see how
this project ever worked for Rails apps with just the sass dependency.